+++ src/bot/templates/src/app/scripts/{botName}/{staticTabClassName}Tab.tsx
@@ -28,7 +28,7 @@ export class <%=staticTabClassName%>Tab extends TeamsBaseComponent<I<%=staticTab
         if (await this.inTeams()) {
             microsoftTeams.initialize();
             microsoftTeams.registerOnThemeChangeHandler(this.updateTheme);
-            microsoftTeams.appInitialization.notifyAppLoaded();
+            microsoftTeams.appInitialization.notifySuccess();
         }
     }